pyproject.toml
[tool.poetry]
name = "bitwise"
version = "0.1.0"
description = ""
authors = ["Pang Hua Yen <panghua.yen@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.10,<3.12"
discord = "^2.3.2"
python-dotenv = "^1.0.0"
black = "^23.9.1"
pynacl = "^1.5.0"
jishaku = "^2.5.1"
uvloop = "^0.17.0"
deemix = "^3.6.6"
soundcloud-v2 = "^1.3.1"
selenium-wire = "^5.1.0"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
